(snapshot) RISCV: Added argument passing test case.
An argument-passing test case was implemented to analyze
how arguments are passed in a function call—whether
through registers, the stack, or a combination of both.

This program is expected to dump all register values, and
stack up to 128 bytes after a function call.

Currently for RISCV 32 and 64 bit only.

Signed-off-by: Luis Silva <luiss@synopsys.com>