(snapshot) [argPassTests] Update in finding magic numbers logic
This revised approach considers the hexadecimal representation
of dump values for compiler registers and stacks up to 128
within a function call.

It will firstly, iterate over a list of magic numbers associated
with a given datatype. For each magic number, it converts it to
hexadecimal and searches for a match in a provided mapping.
If a match is found, it stores the associated key and adds it
to a list of found magic numbers. If no complete match is found,
it splits the hexadecimal value into upper and lower parts and
attempts to find matches for each part separately.
Finally, it generates warnings for any duplicate keys found for
the same hexadecimal value and returns the list of found
magic numbers along with any warnings.

Signed-off-by: Luis Silva <luiss@synopsys.com>