[datatypeTests] Add Signedness Test Case
A signedness test case has been added to
verify if a given data type supports negative numbers.

The test assigns a value of "-1" to a temporary
variable of the data type and then checks its value.
If the value remains negative, it indicates that
the data type supports negative numbers.

Signed-off-by: Luis Silva <luiss@synopsys.com>