(snapshot) [stack_dir] Add stack direction test case.
This test case is designed to determine the stack
direction by using two function calls. Function "A"
defines a local variable and tracks its address in a
global variable. Function "A" then calls "B", where
"B" also defines a local variable and tracks its
address in another global variable.

This approach allows to determine the stack
direction: if the nested stack address is higher,
the stack grows upwards; if it is lower, the stack
grows downwards.

Signed-off-by: Luis Silva <luiss@synopsys.com>