(snapshot) [stackAlign] Add stack alignment test case.
This test case calculates the alignment of the stack pointer at each
recursive call.
Each function TrackAlignment1 to TrackAlignment64 allocates a local array
of increasing size, from 1 to 64 bytes, simulating stack growth with
recursive function calls.

The value of the stack pointer after each function call is tracked using
an assembly function implementation, and stored in.

To determine the stack alignment in the main function, we count the
number of trailing zero bits in the tracked variable. This count represents
the number of bits the stack pointer is aligned to, which is equivalent
to the stack alignment in bytes.

Signed-off-by: Luis Silva <luiss@synopsys.com>