[datatypesTests] Update union alignment and size test case.
Updated the test case to validate union type alignment
and size for the following data types:
    "char"    "signed char"    "unsigned char"    "short"
    "int"     "long"           "long long"        "void*"
    "float"   "double"	       "long double"

Given the 32-bit example:

union Example {
	char a;    // 1-byte alignment
	int b;     // 4-byte alignment
};

The memory layout will look like this:
+---------------+-------+-----------+-----+
| Byte Offset	|   0   |  1  |  2  |  3  |
| Data		|  a/b  |  b  |  b  |  b  |
+---------------+-------+-----+-----+-----+
	- `a` occupies 1 byte;
	- `b`: occupies 4 bytes (the largest member).

The alignment requirement of the union is 4 bytes (the strictest
alignment among its members).
The size of the union is 4 bytes, which is a multiple of its
alignment (4 bytes).

------

TODO: This test case needs improvement by mixing different
data types in the same struct to validate the following conditions:

RISC-V ABIs Specification v1.1
 * Chapter 4. C/C++ type details
 ** 4.1. C/C++ type sizes and alignments
 ***  "Structs and unions are aligned to the alignment of their most strictly
 ***   aligned number. The size of any object is a multiple of its alignment"

Signed-off-by: Luis Silva <luiss@synopsys.com>