(snapshot) [emptyStructTests] Add validator for empty struct test case.
The logic relies on validating the presence of the keyword ("0xdead") in the
argument passing registers for a given architecture to ensure that an empty
struct is ignored by the C compiler.

Here's how it works:
We will pass the empty struct as an argument alongside a sentinel value
denoting the boundaries. Then test the empty struct in all possible positions
in the argument list, for example:

callee (S, I)
callee (I, S, I)
callee (I, I, S, I)
...

where:
    "S" represents the empty struct
    "I" represents the sentinel value.

For each information dump (each "callee" call), it is verified that only the
value of "I" ("0xdead") is present in the argument passing registers. This
confirms that the empty struct is indeed being ignored by the C compiler.

Signed-off-by: Luis Silva <luiss@synopsys.com>