(snapshot) [structGen] Add generator for struct argument passing test case.
The purpose of this generator is to create a test case
that will validate the boundaries before a struct
containing values goes to the stack.

It will generate a test case for a given datatype with
multiple structs being passed to a extern callee() function
responsible for dumping the register and stack values.

These structs are created from 1 to a given MaxCount, where
the last MaxCount will always contain one last "char" element
to validate the boundaries.

Signed-off-by: Luis Silva <luiss@synopsys.com>