(snapshot) [abi-extract-info] Add configuration for struct argument passing test case.
Added the initial configuration for the struct argument
passing test case where it will validate, for each datatype,
the size boundaries of structs before they are passed to the
stack.

Currently, it validates it by verifying when the struct is
passed by reference with the stack address as the first
argument register. This might need change as some
architectues could have different behaviours.

Signed-off-by: Luis Silva <luiss@synopsys.com>