(snapshot) [src/helper.c] Fixing duplicate hex identifiers in C output.
When printing a pointer in C, the hexadecimal identifier "0x" is
already included, so there is no need to prepend another "0x". Adding
an extra "0x" causes issues with value comparison after parsing.

Additionally, the output of "sizeof(Stack)" had a redundant "0x",
which has been corrected to a single "0x".

Output before the patch:
    // Header info
    0x0x8001ffd0
    0x0x00000004
    (...)

Output after the patch:
    // Header info
    0x8001ffd0
    0x00000004
    (...)

Signed-off-by: Luis Silva <luiss@synopsys.com>