(snapshot) [lib/targetArch] Store datatype details value
Some test cases will rely significantly on the
characteristics of others. Specifically, they might
consider the signedness, size, and alignment for
each datatype.

These details are stored in the `targetArch` class,
which manages architecture-specific information.

As these details are common across all architectures,
they are stored in the parent class.

Signed-off-by: Luis Silva <luiss@synopsys.com>