(snapshot) [abi-extract-info] Adjusting struct size boundary limits
For RISCV 32-bit, if a struct exceeds 16 bytes in size for argument
passing, the reference address in `a0` will not correspond to the
stack pointer but will include an offset, causing issues with the
current logic.
TODO: Test for 64-bit scenarios.

Initially, the `max_boundary` was set to twice the size of an `int`,
based on the RISCV ABI document. However, this approach is adjusted
to use the size of a `char` to define the limits, as the previous
assumption may not apply to other architectures. The size of an `int`
is still used to finalize the `max_boundary` value.

Signed-off-by: Luis Silva <luiss@synopsys.com>