(snapshot) [lib/argPassTestGen] Restructure argument passing test case gen.
The purpose of this class is to generate a variety of C test cases using
different fundamental C types for argument passing.

Given a specified argument count value, several hexadecimal values will be
passed to a `callee` function written in assembly, which saves all register
values and prints them to stdout at the end.

For instance, using `int` as the datatype and 4 as argument count,
we have the following example:
```c
    extern void callee(int, int, int, int);

    int main(void) {
        callee(0x12345678, 0x12345678, 0x12345678, 0x12345678);
    }
```

Note that for certain datatypes, an auxiliary function is created to convert
the value of one datatype to another. For example, for the `double` datatype,
the test case is as follows:
```c
    #include <string.h>

    inline static double ull_as_double(unsigned long long lhs) {
        double result;
        memcpy(&result, &lhs, sizeof(result));
        return result;
    }

    extern void callee(double, double, double, double);

    int main(void) {
        callee(ull_as_double(0x1234567890abcdef), ull_as_double(0x1234567890abcdef),
            ull_as_double(0x1234567890abcdef), ull_as_double(0x1234567890abcdef));
    }
```
Hexadecimal values like `0x1234567890abcdef` are typically interpreted as
integer types in C, such as `unsigned long long`.

When dealing with different datatypes like `double`, we need to be aware that
although both `unsigned long long` and `double` may occupy the same number
of bits (64 bits), their bit-level representations are different.

The convertion function `ull_as_double` uses `memcpy` to directly copy the
bit pattern of the `unsigned long long` into a `double`. This does not change
the bit values themselves but reinterprets them according to the
IEEE 754 standard for a `double`.

Without this convertion function, the compiler would not know to correctly
interpret the 64-bit pattern as a `double` and will likely change the
hexadecimal value.

Signed-off-by: Luis Silva <luiss@synopsys.com>