(snapshot) [lib/argPassTests] Find registers from value or split halves.
If the target value is not found in any of the registers, this value
will be splitten in two, and then compare each half to the register
values. This is done because, for example, a `double` value might
not fit in a single 32-bit register. So, the hexadecimal value is
splitten into two registers.

Signed-off-by: Luis Silva <luiss@synopsys.com>