(snapshot) [lib/argPassTests] Find registers that hold split halves.
After splitting a hexadecimal value in half, this function will
compare the halves to the register dump values. If those are found,
those are saved while validating the order of the halves values
in the registers. Being either [Low] & [High] or [High] & [Low] by
comparing the first register value found to the first half.

Signed-off-by: Luis Silva <luiss@synopsys.com>