(snapshot) [lib/argPassTests] Extract common registers in list of list registers.
For example, given a N number of iterations of N argument
passing. Each will provide a set of registers used in the
passing of arguments. The idea behind this function is to
detect inconsistencies in the registers used in all the
iterations.

If one register is used in one iteration, but not in others,
it might imply a inconsistency and will be returned as a
non-common register. Otherwise, it will be returned as a
common register list.

Signed-off-by: Luis Silva <luiss@synopsys.com>