(snapshot) [src/arch/riscv.s] Fix return address.
The return address was being saved in the `t0` register,
which was not the correct approach as that register may
be modified in the `dump_information` function.

This approach has been changed to save the return address
to a callee-saved register `s2`, so that after the
`dump_information` function call, it can be restored
successfully.

Signed-off-by: Luis Silva <luiss@synopsys.com>