(snapshot) [hexUtils] Extend value to argc.
Extend the given hexadecimal value to the specified argument count
and return list of hexadecimal values.

This ensures if multiple datatypes are being passed in the same
chunk of memory. For example, given an argument count of 3 with
`char` as datatypes, they can be passed in one single register.

    Char values: { 0x12, 0x12, 0x12 }
    Register: 0x121212

Another validation is if we have a register size of 4 bytes, and
the extended value is greater, we need to split the values into
two (or more) chunks, and search them in the register banks.

    Value        : 0x1212121212
    First chunk  : 0x12121212
    Second chunk : 0x12

Signed-off-by: Luis Silva <luiss@synopsys.com>