(snapshot) [lib/returnGen] Add return test case generator.
The purpose of this generator is to create a test case
that will validate how datatype values are passed in
return calls.

It will create a call to an assembly function, which
will then call a C function that will return a
datatype value, that then all registers and stack
is dumped.

             main    (.c)
              v
             foo     (.s)
              v
             bar     (.c)
              v
            callee   (.s)

Signed-off-by: Luis Silva <luiss@synopsys.com>