[lib/saveGen] Prevent compiler optimization.
Some compilers might optimize the `aux` function and
skip it entirely. To prevent that, seperate files for
`main` and `aux` functions have been created.

Another optmization that might happen is related to the
handler of the `set_registers` function call. The compiler
might restore all registers to be able to use the return address
registers to jump directly from the `set_registers` to `main`,
instead of jumping back to `aux` and then `main`. To prevent this,
we return an address of a function from `aux` to `main`.

Signed-off-by: Luis Silva <luiss@synopsys.com>