(snapshot) [src/arch/riscv] Update file suffix.
In GCC, there is a different between an assembly file
that ends with .s and .S.

With .S the driver calls the preprocessor before the
assembler, this way it is possible to include macros
in the assembly code.

With .s the driver does not call the preprocessor at all,
leaving with an error of an undefined constant.

Signed-off-by: Luis Silva <luiss@synopsys.com>