(snapshot) [lib/argPassTests] Update run test wrapper.
In the RISCV ilp32d ABI, floating-point registers are 64 bits
in size, even though the architecture itself is 32-bit. When passing
a `double` (which is 64 bits), it uses the floating-point registers
`fpa0` through `fpa7`, which can hold the entire value. Once these
floating-point registers are used, the calling convention switches
to `a0` through `a7` registers, which only 32 bits each. Consequently,
while the `fpa0` through `fpa7` registers store the value in its entirety,
the `a0` through `a7` registers require the value to be split into
low and high parts.

Previously, if a value was found in its entirety in the dump
information, there was no further search for split values. However,
with the situation described, it is necessary to also validate split
values.

Signed-off-by: Luis Silva <luiss@synopsys.com>