(snapshot) [lib/hexUtils] Fix inconsistencies list.
In this case, we are only validating the last value
in the argument list, which means there can only be
a single value with inconsistencies. As a result,
there is no need to use a temporary list to store
multiple inconsistencies; we can store it directly
in the inconsistencies list.

Signed-off-by: Luis Silva <luiss@synopsys.com>