(snapshot) [lib/bitFieldGen] Simplify struct and union generation.
Instead of using seperated struct and unions, this patch
combines the struct inside the union.

The new `values` array will be mostly used because of values
that wont fit in a single long long variable.

For example, when dealing with `long long` datatype as bitfields,
if the sum of these bitfields exceeds the sizeof `long long`, each
value will fit in a `long long`. To retrieve the whole values, we need
that array.

Signed-off-by: Luis Silva <luiss@synopsys.com>