(snapshot) [lib/bitFieldGen] Add support for `long long` bitfields.
`long long` bitfields behave the same as any other bitfield. If the
combined bitfieds can fit inside a single datatype, it will. If not,
each value will occupy a seperate datatype.

The problem with `long long` its its size. So we need to use lower
and upper variables to hold the respective value.

Signed-off-by: Luis Silva <luiss@synopsys.com>