(snapshot) [lib/bitFieldGen] Generate bitfield examples.
For each datatype we need a set of examples to make sure
we have enough sample to generate the report. One example
for each (greater and less than sizeof datatype) might not
enough because there could be cases where there can be
multiple results for the same value.

Currently we have 6 for each datatype (3 greater than and 3 less than).

Signed-off-by: Luis Silva <luiss@synopsys.com>