[lib/helper] Add reset option to binary gen.
Once a binary value is generated it gets stored so that
that value is not duplicated later on. The problem arises
with low sized values.

For example, a sizeof of 1 bit can only have one outcome,
which is 1. If 1 gets stored to the `used_values` array, it
can no longer be reused. Well, if another 1 bit value needs
to be generated, it will be stuck in an infinite loop.

To fix this, a reset option to the `used_values` array has
been added.

Signed-off-by: Luis Silva <luiss@synopsys.com>