(snapshot) [abi-extractor-info] Generate caller/callee values.
For the caller/callee saved registers test case,
we need to generate values for both the C test case
creation and the validation of those values.

Rather than generating them at multiple points
(savedGen.py and savedTest.py) with the same seed,
they are now generated once in this location.

Signed-off-by: Luis Silva <luiss@synopsys.com>