(snapshot) [lib/structGen] Add struct size dump.
Print the struct's size to check the size boundaries
based on its contained elements. For example, we cannot
assume that a struct containing a single char will always
be one byte, as this can vary depending on the
architecture and ABI.

Also the size can vary depending on the alignment
requirements of different data types within the struct.

Signed-off-by: Luis Silva <luiss@synopsys.com>