Create a base class for analyzer implementations
The bass class deals with handling temporary files, compilation, and
running (simulated) binaries. These steps should be identical for each
analyzer, so this should eliminate some potential mistakes.

The analyzer just has to specify how to generate source code for the
test and how to handle stdout. This should make testing easier when all
IO is refactored outside of the analyzer itself.

Signed-off-by: Michiel Derhaeg <michiel@synopsys.com>