Remove reliance on passing CONFIG_FPU in the wrappers
Passing this definition in your custom wrapper is easy to forget and
causes hard-to-debug errors.

Instead we check for `__riscv_float_abi_soft`. If this is predefined
hardware floats are not supported by the ABI.

Also fix a minor bug in the struct boundaries analyzer. It only
triggered when hardware floats were not supported.

Signed-off-by: Michiel Derhaeg <michiel@synopsys.com>